# specs/001-agent-service-cpq/contracts/openapi.template.yaml
openapi: 3.1.0
info:
  title: ClientWeave Public and Agent API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Generated external contract for the ordinary buyer UI and six page-provided
    agent capabilities. Authoritative executable schemas live under
    src/contracts/schemas/. Owner administration uses the same domain services
    behind authenticated server actions and is described by the plan and data model.
servers:
  - url: /api/v1
...
paths:
...
  /scopes:
    post:
      tags: [Scope]
      operationId: create_scope
      summary: Create one durable shared draft scope
      security: []
      x-agent-capability:
        stateEffect: DRAFT_MUTATION
        consequential: false
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyOptional'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateScopeRequest' }
      responses:
        '201':
          description: Scope created; continuationUrl contains a one-time bearer token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ScopeReview'
                  - type: object
                    required: [continuationUrl]
        '400': { $ref: '#/components/responses/ValidationError' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    scopeBearer:
      type: http
      scheme: bearer
      bearerFormat: OpaqueScopeToken
      description: Agent credential authorizing only one unexpired scope; stored hashed server-side.
    scopeSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-clientweave_scope
      description: Secure HttpOnly browser session authorizing only one unexpired scope.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with the scope session cookie for browser mutations.
  parameters:
...
    IdempotencyKeyOptional:
      in: header
      name: Idempotency-Key
      required: false
      schema: { type: string, minLength: 16, maxLength: 128 }
  responses:
...
    ValidationError:
      description: Input or current-state validation failed safely
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Conflict:
      description: Revision, quote, confirmation, idempotency, or slot conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    RateLimited:
      description: Public request limit exceeded
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
    MoneyRange:
      type: object
      required: [currency, minMinor, maxMinor]
      properties:
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        minMinor: { type: integer, minimum: 0 }
        maxMinor: { type: integer, minimum: 0 }
      additionalProperties: false
    ServiceMatch:
      type: object
      required: [serviceId, slug, name, description, basePrice, deliveryBusinessDays, includedItems, addOns, intakeFields, constraints, fit]
      properties:
        serviceId: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        basePrice: { $ref: '#/components/schemas/MoneyRange' }
        deliveryBusinessDays:
          type: object
          required: [min, max]
          properties:
            min: { type: integer, minimum: 1 }
            max: { type: integer, minimum: 1 }
          additionalProperties: false
        includedItems: { type: array, items: { type: string } }
        addOns: { type: array, items: { type: string } }
        intakeFields:
          type: array
          items: { $ref: '#/components/schemas/ScopeField' }
        constraints:
          type: array
          items: { $ref: '#/components/schemas/ServiceConstraintSummary' }
        fit:
          type: object
          required: [eligible, reasons, conflicts]
          properties:
            eligible: { type: boolean }
            reasons: { type: array, items: { type: string } }
            conflicts: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } }
          additionalProperties: false
      additionalProperties: false
    ScopeField:
      type: object
      required: [key, label, valueType, required, priceAffecting]
      properties:
        key: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        label: { type: string }
        helpText: { type: string }
        valueType: { enum: [TEXT, INTEGER, DECIMAL, BOOLEAN, DATE, SINGLE_CHOICE, MULTI_CHOICE] }
        required: { type: boolean }
        priceAffecting: { type: boolean }
        allowedValues: { type: array, items: { type: [string, number, boolean] } }
      additionalProperties: false
    CreateScopeRequest:
      type: object
      required: [serviceId, goal]
      properties:
        serviceId: { type: string, format: uuid }
        goal: { type: string, minLength: 1, maxLength: 2000 }
        budgetMaxMinor: { type: integer, minimum: 0 }
        targetDeliveryDate: { type: string, format: date }
        assumptions:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 500 }
        answers:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/AnswerInput' }
      additionalProperties: false
...
    AnswerInput:
      type: object
      required: [fieldKey, value]
      properties:
        fieldKey: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        value: {}
      additionalProperties: false
    ScopeReview:
      type: object
      required: [scopeRef, status, revision, service, goal, budgetMaxMinor, targetDeliveryDate, assumptions, answers, missingFields, conflicts, quoteStatus, expiresAt]
      properties:
        scopeRef: { type: string, minLength: 16, maxLength: 128 }
        status: { enum: [DRAFT, FINALIZED, EXPIRED] }
        revision: { type: integer, minimum: 1 }
        service: { $ref: '#/components/schemas/ServiceMatch' }
        goal: { $ref: '#/components/schemas/AttributedTextValue' }
        budgetMaxMinor: { $ref: '#/components/schemas/AttributedNullableMoneyValue' }
        targetDeliveryDate: { $ref: '#/components/schemas/AttributedNullableDateValue' }
        assumptions:
          type: array
          maxItems: 20
          items: { $ref: '#/components/schemas/AttributedTextValue' }
        answers:
          type: array
          items: { $ref: '#/components/schemas/AttributedAnswer' }
        missingFields: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } }
        conflicts: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } }
        quoteStatus: { enum: [NONE, CURRENT, STALE, INCOMPLETE, CONFLICTED] }
        expiresAt: { type: string, format: date-time }
        continuationUrl:
          type: string
          format: uri
          description: Returned only once from scope creation; its secret is in the URL fragment.
      additionalProperties: false
    ValidationIssue:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        fieldKey: { type: string }
        message: { type: string }
        severity: { enum: [INFO, WARNING, ERROR], default: ERROR }
        suggestedActions: { type: array, items: { type: string } }
      additionalProperties: false
    ServiceConstraintSummary:
      type: object
      required: [type, fieldKeys, message]
      properties:
        type: { enum: [DELIVERY_MINIMUM, INCOMPATIBLE_OPTIONS, FIELD_LIMIT, BUDGET_FLOOR] }
        fieldKeys:
          type: array
          items: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        message: { type: string, minLength: 1, maxLength: 500 }
      additionalProperties: false
    ProvenanceMetadata:
      type: object
      required: [actorType, updatedAt]
      properties:
        actorType: { enum: [HUMAN, AGENT, IMPORTED, SYSTEM] }
        actorId: { type: [string, 'null'], maxLength: 255 }
        updatedAt: { type: string, format: date-time }
      additionalProperties: false
    AttributedTextValue:
      type: object
      required: [value, provenance]
      properties:
        value: { type: string }
        provenance: { $ref: '#/components/schemas/ProvenanceMetadata' }
      additionalProperties: false
    AttributedNullableMoneyValue:
      type: object
      required: [value, provenance]
      properties:
        value: { type: [integer, 'null'], minimum: 0 }
        provenance: { $ref: '#/components/schemas/ProvenanceMetadata' }
      additionalProperties: false
    AttributedNullableDateValue:
      type: object
      required: [value, provenance]
      properties:
        value: { type: [string, 'null'], format: date }
        provenance: { $ref: '#/components/schemas/ProvenanceMetadata' }
      additionalProperties: false
    AttributedAnswer:
      type: object
      required: [fieldKey, value, provenance]
      properties:
        fieldKey: { type: string, pattern: '^[a-z][a-z0-9_]{0,63}$' }
        value: {}
        provenance: { $ref: '#/components/schemas/ProvenanceMetadata' }
      additionalProperties: false
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail, code, retryable, traceId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
        retryable: { type: boolean }
        fieldErrors: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } }
        currentRevision: { type: integer, minimum: 1 }
        traceId: { type: string }